+++ botfarm/src/main/java/io/deeplay/camp/botfarm/bots/denis_bots/RandomUtility.java
@@ -1,7 +1,6 @@
 package io.deeplay.camp.botfarm.bots.denis_bots;
 
-import io.deeplay.camp.botfarm.bots.Bot;
-import io.deeplay.camp.game.entities.UnitType;
+import io.deeplay.camp.botfarm.bots.denis_bots.entities.BotTactic;
 import io.deeplay.camp.game.events.MakeMoveEvent;
 import io.deeplay.camp.game.events.PlaceUnitEvent;
 import io.deeplay.camp.game.mechanics.GameState;
@@ -10,6 +9,10 @@ import io.deeplay.camp.game.mechanics.PlayerType;
 import java.util.List;
 import java.util.Random;
 
+/**
+ * Класс реализующий все вспомогательные общие функции для бота, но с рандомными
+ * значениями.
+ */
 public class RandomUtility implements UtilityFunction{
     BotTactic botTactic;
     PlayerType currentPlayerType;
