+++ botfarm/src/main/java/io/deeplay/camp/botfarm/bots/denis_bots/entities/BotTactic.java
@@ -1,5 +1,9 @@
-package io.deeplay.camp.botfarm.bots.denis_bots;
+package io.deeplay.camp.botfarm.bots.denis_bots.entities;
 
+/**
+ * Энам, реализующий тактики для бота, определяющая его
+ * Генерала.
+ */
 public enum BotTactic {
     MAGE_TACTIC,
     KNIGHT_TACTIC,
