+++ botfarm/src/main/java/io/deeplay/camp/botfarm/bots/denis_bots/movement_algorithm/NewClastExpMaxAlg.java
@@ -1,61 +1,33 @@
-package io.deeplay.camp.botfarm.bots.denis_bots;
+package io.deeplay.camp.botfarm.bots.denis_bots.movement_algorithm;
 
-import io.deeplay.camp.botfarm.bots.Bot;
-import io.deeplay.camp.botfarm.bots.RandomBot;
-import io.deeplay.camp.game.entities.*;
+import io.deeplay.camp.botfarm.bots.denis_bots.tools.KMeans;
+import io.deeplay.camp.botfarm.bots.denis_bots.UtilityFunction;
+import io.deeplay.camp.botfarm.bots.denis_bots.entities.UtilityMoveResult;
+import io.deeplay.camp.game.entities.StateChance;
 import io.deeplay.camp.game.events.MakeMoveEvent;
-import io.deeplay.camp.game.events.PlaceUnitEvent;
 import io.deeplay.camp.game.exceptions.GameException;
 import io.deeplay.camp.game.mechanics.GameStage;
 import io.deeplay.camp.game.mechanics.GameState;
-import io.deeplay.camp.game.mechanics.PlayerType;
-import lombok.Setter;
 import lombok.SneakyThrows;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.List;
 import java.util.ArrayList;
-
+import java.util.Collections;
+import java.util.List;
 import java.util.concurrent.ForkJoinTask;
 import java.util.concurrent.RecursiveTask;
 
-public class ModNewClastExpMaxBot extends Bot {
-    private static final Logger logger = LoggerFactory.getLogger(RandomBot.class);
-    BotTactic botTactic;
+public class NewClastExpMaxAlg {
+    final int originDepth;
     UtilityFunction tacticUtility;
-    UnitType currentGeneral;
-    int maxDepth;
-    @Setter boolean firstPlaceInGame = true;
     double eps = 0.001;
-
-    public ModNewClastExpMaxBot(PlayerType playerType, int maxDepth){
-        tacticUtility = new TacticUtility(BotTactic.BASE_TACTIC);
-        tacticUtility.setCurrentPlayerType(playerType);
-        this.maxDepth = maxDepth;
+    public NewClastExpMaxAlg(int maxDepth, UtilityFunction tacticUtility){
+        this.originDepth = maxDepth;
+        this.tacticUtility = tacticUtility;
     }
 
 
-    @Override
-    public PlaceUnitEvent generatePlaceUnitEvent(GameState gameState) {
-        List<PlaceUnitEvent> placeUnitEvents = gameState.getPossiblePlaces();
-        if(!placeUnitEvents.isEmpty()){
-            return placeUnitEvents.get((int)(Math.random()*placeUnitEvents.size()));
-        }
-        else{
-            return null;
-        }
-    }
-
-    @Override
-    public MakeMoveEvent generateMakeMoveEvent(GameState gameState) {
-        firstPlaceInGame = true;
-
-        return getMoveResult(gameState).event;
-    }
-
-    private UtilityMoveResult getMoveResult(GameState gameState) {
-        int originDepth = maxDepth;
+    public UtilityMoveResult getMoveResult(GameState gameState) {
+        int originDepth = this.originDepth;
         List<MakeMoveEvent> movesRoot = gameState.getPossibleMoves();
 
         if (movesRoot.isEmpty()) {
@@ -66,28 +38,26 @@ public class ModNewClastExpMaxBot extends Bot {
                 UtilityMoveResult features = extractValue(gameState, move);
                 points.add(features);
             }
+
             KMeans kMeans = new KMeans();
             int numClusters = Math.min(5, movesRoot.size());  // количество кластеров можно варьировать
             List<KMeans.Cluster> clusters = kMeans.kMeansCluster(points, numClusters, 100);
 
-
-            List<UtilityMoveResult> bestMoves = new ArrayList<>();
-
-
+            List<UtilityMoveResult> bestMoves;
 
             UtilityMoveResult bestResult = new UtilityMoveResult(Double.NEGATIVE_INFINITY, null);
             KMeans.Cluster bestCluster = null;
             for (KMeans.Cluster cluster : clusters) {
                 UtilityMoveResult bestMoveInCluster = getBestMoveInCluster(points ,cluster, movesRoot, gameState, true);
-                if(bestMoveInCluster.event != null) {
-                    if(bestMoveInCluster.value > bestResult.value){
+                if(bestMoveInCluster.getEvent() != null) {
+                    if(bestMoveInCluster.getValue() > bestResult.getValue()){
                         bestCluster = cluster;
                         bestResult = bestMoveInCluster;
                     }
-                    //bestMoves.add(bestMoveInCluster);
                 }
             }
             bestMoves = bestCluster.points;
+            bestMoves.sort(Collections.reverseOrder());
 
 
 
@@ -97,22 +67,23 @@ public class ModNewClastExpMaxBot extends Bot {
                     @SneakyThrows
                     @Override
                     protected UtilityMoveResult compute() {
-                        double result = expectMaxAlg(gameState, moveBestClast.event, originDepth, true, 1);
-                        return new UtilityMoveResult(result, moveBestClast.event);
+                        double result = expectMaxAlg(gameState, moveBestClast.getEvent(), originDepth, true, 1);
+                        return new UtilityMoveResult(result, moveBestClast.getEvent());
                     }
                 };
                 recursiveTasks.add(task);
 
             }
+
             List<UtilityMoveResult> results = ForkJoinTask.invokeAll(recursiveTasks).stream()
                     .map(ForkJoinTask::join)
                     .toList();
 
-            return getMaxFromTasks(results);
+            return getMaxMoveFromTasks(results);
         }
     }
 
-    public double minMaxAlg(GameState root, int depth, boolean maxPlayer, double p) {
+    private double minMaxAlg(GameState root, int depth, boolean maxPlayer, double p) {
         if(depth == 0 || root.getGameStage() == GameStage.ENDED || p < eps){
             return tacticUtility.getMoveUtility(root);
         }
@@ -123,28 +94,26 @@ public class ModNewClastExpMaxBot extends Bot {
             UtilityMoveResult features = extractValue(root, move);
             points.add(features);
         }
+
         KMeans kMeans = new KMeans();
         int numClusters = Math.min(5, movesRoot.size());  // количество кластеров можно варьировать
         List<KMeans.Cluster> clusters = kMeans.kMeansCluster(points, numClusters, 100);
 
         List<UtilityMoveResult> bestMoves = new ArrayList<>();
 
-
         UtilityMoveResult bestResult = new UtilityMoveResult(Double.NEGATIVE_INFINITY, null);
         KMeans.Cluster bestCluster = null;
         for (KMeans.Cluster cluster : clusters) {
             UtilityMoveResult bestMoveInCluster = getBestMoveInCluster(points ,cluster, movesRoot, root, maxPlayer);
-            if(bestMoveInCluster.event != null) {
-                if(bestMoveInCluster.value > bestResult.value){
+            if(bestMoveInCluster.getEvent() != null) {
+                if(bestMoveInCluster.getValue() > bestResult.getValue()){
                     bestCluster = cluster;
                     bestResult = bestMoveInCluster;
                     bestMoves = bestCluster.points;
                 }
-                //bestMoves.add(bestMoveInCluster);
             }
         }
-
-
+        bestMoves.sort(Collections.reverseOrder());
 
 
         if(movesRoot.isEmpty()){
@@ -163,7 +132,7 @@ public class ModNewClastExpMaxBot extends Bot {
                     @Override
                     protected UtilityMoveResult compute() {
                         try {
-                            return new UtilityMoveResult(expectMaxAlg(root, moveEvent.event, depth, maxPlayer, p), moveEvent.event);
+                            return new UtilityMoveResult(expectMaxAlg(root, moveEvent.getEvent(), depth, maxPlayer, p), moveEvent.getEvent());
                         } catch (GameException e) {
                             throw new RuntimeException(e);
                         }
@@ -175,7 +144,7 @@ public class ModNewClastExpMaxBot extends Bot {
                     .map(ForkJoinTask::join)
                     .toList();
 
-            return maxPlayer ? getMaxFromTasks(results).getValue() : getMinFromTasks(results).getValue();
+            return maxPlayer ? getMaxMoveFromTasks(results).getValue() : getMinMoveFromTasks(results).getValue();
         }
     }
 
@@ -200,10 +169,10 @@ public class ModNewClastExpMaxBot extends Bot {
         return new UtilityMoveResult(tacticUtility.getMoveUtility(gameStateNode), move);
     }
 
-    private UtilityMoveResult getMaxFromTasks(List<UtilityMoveResult> results){
+    private UtilityMoveResult getMaxMoveFromTasks(List<UtilityMoveResult> results){
         UtilityMoveResult bestValue = new UtilityMoveResult(Double.NEGATIVE_INFINITY, null);
         for (UtilityMoveResult task : results) {
-            if (bestValue.value < task.value) {
+            if (bestValue.getValue() < task.getValue()) {
                 bestValue = task;
             }
         }
@@ -211,25 +180,16 @@ public class ModNewClastExpMaxBot extends Bot {
         return bestValue;
     }
 
-    private UtilityMoveResult getMinFromTasks(List<UtilityMoveResult> results){
+    private UtilityMoveResult getMinMoveFromTasks(List<UtilityMoveResult> results){
         UtilityMoveResult bestValue = new UtilityMoveResult(Double.POSITIVE_INFINITY, null);
         for (UtilityMoveResult task : results) {
-            if (task.value < bestValue.value) {
+            if (task.getValue() < bestValue.getValue()) {
                 bestValue = task;
             }
         }
         return bestValue;
     }
 
-    public List<Position> enumerationPlayerUnits(PlayerType playerType, Board board) {
-        List<Position> unitPositions = new ArrayList<>();
-        if (playerType == PlayerType.FIRST_PLAYER) {
-            unitPositions.addAll(board.enumerateUnits(0, Board.ROWS / 2));
-        } else {
-            unitPositions.addAll(board.enumerateUnits(Board.ROWS / 2, Board.ROWS));
-        }
-        return unitPositions;
-    }
     @SneakyThrows
     private UtilityMoveResult getBestMoveInCluster(List<UtilityMoveResult> points, KMeans.Cluster cluster, List<MakeMoveEvent> movesRoot, GameState gameState, boolean maxPlayer) {
         List<RecursiveTask<UtilityMoveResult>> tasks = new ArrayList<>();
@@ -250,6 +210,7 @@ public class ModNewClastExpMaxBot extends Bot {
                 .map(ForkJoinTask::join)
                 .toList();
 
-        return maxPlayer ? getMaxFromTasks(results) : getMinFromTasks(results);
+        return maxPlayer ? getMaxMoveFromTasks(results) : getMinMoveFromTasks(results);
     }
+
 }
