+++ botfarm/src/main/java/io/deeplay/camp/botfarm/bots/denis_bots/ready_bots/ModNewClastMetricPlaceExpMaxBot.java
@@ -1,0 +1,59 @@
+package io.deeplay.camp.botfarm.bots.denis_bots.ready_bots;
+
+import io.deeplay.camp.botfarm.bots.Bot;
+import io.deeplay.camp.botfarm.bots.RandomBot;
+import io.deeplay.camp.botfarm.bots.denis_bots.entities.BotTactic;
+import io.deeplay.camp.botfarm.bots.denis_bots.TacticUtility;
+import io.deeplay.camp.botfarm.bots.denis_bots.UtilityFunction;
+import io.deeplay.camp.botfarm.bots.denis_bots.movement_algorithm.NewClastExpMaxAlg;
+import io.deeplay.camp.botfarm.bots.denis_bots.placement_algorithm.MetricPlaceAlg;
+import io.deeplay.camp.game.events.MakeMoveEvent;
+import io.deeplay.camp.game.events.PlaceUnitEvent;
+import io.deeplay.camp.game.mechanics.GameState;
+import io.deeplay.camp.game.mechanics.PlayerType;
+import lombok.Setter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+
+/**
+ * Бот, использующий алгоритм ЭкспектиМакс, использующий:
+ * Многопоточность,
+ * Улучшенную кластеризацию,
+ * Отсечение по вероятностям,
+ * Расстановку основанную на метрических данных.
+ */
+public class ModNewClastMetricPlaceExpMaxBot extends Bot {
+    private static final Logger logger = LoggerFactory.getLogger(RandomBot.class);
+    UtilityFunction tacticUtility;
+    int maxDepth;
+    @Setter boolean firstPlaceInGame = true;
+    NewClastExpMaxAlg movementAlg;
+    MetricPlaceAlg placementAlg;
+
+    public ModNewClastMetricPlaceExpMaxBot(PlayerType playerType, int maxDepth){
+        tacticUtility = new TacticUtility(BotTactic.KNIGHT_TACTIC);
+        tacticUtility.setCurrentPlayerType(playerType);
+        this.maxDepth = maxDepth;
+        movementAlg = new NewClastExpMaxAlg(maxDepth, tacticUtility);
+        placementAlg = new MetricPlaceAlg(tacticUtility);
+    }
+
+
+    @Override
+    public PlaceUnitEvent generatePlaceUnitEvent(GameState gameState) {
+        if(firstPlaceInGame) {
+            placementAlg.findNewTactic(gameState);
+            firstPlaceInGame = false;
+        }
+        return placementAlg.getPlaceResult(gameState);
+    }
+
+    @Override
+    public MakeMoveEvent generateMakeMoveEvent(GameState gameState) {
+        firstPlaceInGame = true;
+        return movementAlg.getMoveResult(gameState).getEvent();
+    }
+
+
+}
